+++ NEWS.md
@@ -1,6 +1,8 @@
 ## tsfknn 0.3.0
 
 * Now it is possible to assess the model using rolling origin evaluation
+* A predict method has been added to generate new forecasts based on a
+  built model
 
 ## tsfknn 0.2.0